Consolidate TextIOLayer and Dataset.to_file by reusing the same basic kernel ops
