+++ tensorflow_io/core/BUILD
@@ -259,7 +259,6 @@ cc_library(
         "kernels/csv_kernels.cc",
         "kernels/libsvm_kernels.cc",
         "kernels/text_kernels.cc",
-        "kernels/text_layer_kernels.cc",
         "kernels/text_re2.cc",
         "kernels/text_sequence.cc",
         "ops/libsvm_ops.cc",
