+++ tensorflow_io/core/ops/text_ops.cc
@@ -13,10 +13,10 @@ See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
 
+#include "re2/re2.h"
 #include "tensorflow/core/framework/common_shape_fns.h"
 #include "tensorflow/core/framework/op.h"
 #include "tensorflow/core/framework/shape_inference.h"
-#include "re2/re2.h"
 
 namespace tensorflow {
 
@@ -35,35 +35,18 @@ REGISTER_OP("IO>RE2FullMatch")
       TF_RETURN_IF_ERROR(c->GetAttr("pattern", &pattern));
       RE2 re(pattern, RE2::Quiet);
       if (!re.ok()) {
-        return errors::InvalidArgument("unable to compile pattern '", pattern, "': ", re.error());
+        return errors::InvalidArgument("unable to compile pattern '", pattern,
+                                       "': ", re.error());
       }
       shape_inference::ShapeHandle shape;
       TF_RETURN_IF_ERROR(c->WithRankAtLeast(c->input(0), 0, &shape));
       shape_inference::ShapeHandle output_shape;
-      TF_RETURN_IF_ERROR(c->Concatenate(shape, c->Vector(re.NumberOfCapturingGroups()), &output_shape));
+      TF_RETURN_IF_ERROR(c->Concatenate(
+          shape, c->Vector(re.NumberOfCapturingGroups()), &output_shape));
       c->set_output(0, c->input(0));
       c->set_output(1, output_shape);
       return Status::OK();
     });
-
-REGISTER_OP("IO>LayerTextCall")
-  .Input("input: T")
-  .Input("content: string")
-  .Input("resource: resource")
-  .Output("output: T")
-  .Attr("T: type")
-  .SetShapeFn(shape_inference::UnchangedShape);
-
-REGISTER_OP("IO>LayerTextInit")
-  .Input("input: string")
-  .Output("resource: resource")
-  .Attr("container: string = ''")
-  .Attr("shared_name: string = ''")
-  .SetShapeFn(shape_inference::ScalarShape);
-
-REGISTER_OP("IO>LayerTextSync")
-  .Input("resource: resource")
-  .SetShapeFn(shape_inference::ScalarShape);
 
 REGISTER_OP("IO>ReadText")
     .Input("filename: string")
@@ -72,9 +55,9 @@ REGISTER_OP("IO>ReadText")
     .Input("length: int64")
     .Output("output: string")
     .SetShapeFn([](shape_inference::InferenceContext* c) {
-       c->set_output(0, c->MakeShape({c->UnknownDim()}));
-       return Status::OK();
-     });
+      c->set_output(0, c->MakeShape({c->UnknownDim()}));
+      return Status::OK();
+    });
 
 REGISTER_OP("IO>TextOutputSequence")
     .Input("destination: string")
@@ -92,44 +75,44 @@ REGISTER_OP("IO>TextOutputSequenceSetItem")
     .SetShapeFn(shape_inference::ScalarShape);
 
 REGISTER_OP("IO>CSVReadableInit")
-  .Input("input: string")
-  .Output("resource: resource")
-  .Output("components: string")
-  .Attr("container: string = ''")
-  .Attr("shared_name: string = ''")
-  .SetShapeFn([](shape_inference::InferenceContext* c) {
-    c->set_output(0, c->Scalar());
-    c->set_output(1, c->MakeShape({}));
-    return Status::OK();
-   });
+    .Input("input: string")
+    .Output("resource: resource")
+    .Output("components: string")
+    .Attr("container: string = ''")
+    .Attr("shared_name: string = ''")
+    .SetShapeFn([](shape_inference::InferenceContext* c) {
+      c->set_output(0, c->Scalar());
+      c->set_output(1, c->MakeShape({}));
+      return Status::OK();
+    });
 
 REGISTER_OP("IO>CSVReadableSpec")
-  .Input("input: resource")
-  .Output("shape: int64")
-  .Output("dtype: int64")
-  .Attr("component: string")
-  .SetShapeFn([](shape_inference::InferenceContext* c) {
-    c->set_output(0, c->MakeShape({c->UnknownDim()}));
-    c->set_output(1, c->MakeShape({}));
-    return Status::OK();
-   });
+    .Input("input: resource")
+    .Output("shape: int64")
+    .Output("dtype: int64")
+    .Attr("component: string")
+    .SetShapeFn([](shape_inference::InferenceContext* c) {
+      c->set_output(0, c->MakeShape({c->UnknownDim()}));
+      c->set_output(1, c->MakeShape({}));
+      return Status::OK();
+    });
 
 REGISTER_OP("IO>CSVReadableRead")
-  .Input("input: resource")
-  .Input("start: int64")
-  .Input("stop: int64")
-  .Output("value: dtype")
-  .Attr("filter: list(string) = []")
-  .Attr("component: string")
-  .Attr("shape: shape")
-  .Attr("dtype: type")
-  .SetShapeFn([](shape_inference::InferenceContext* c) {
-    PartialTensorShape shape;
-    TF_RETURN_IF_ERROR(c->GetAttr("shape", &shape));
-    shape_inference::ShapeHandle entry;
-    TF_RETURN_IF_ERROR(c->MakeShapeFromPartialTensorShape(shape, &entry));
-    c->set_output(0, entry);
-    return Status::OK();
-   });
+    .Input("input: resource")
+    .Input("start: int64")
+    .Input("stop: int64")
+    .Output("value: dtype")
+    .Attr("filter: list(string) = []")
+    .Attr("component: string")
+    .Attr("shape: shape")
+    .Attr("dtype: type")
+    .SetShapeFn([](shape_inference::InferenceContext* c) {
+      PartialTensorShape shape;
+      TF_RETURN_IF_ERROR(c->GetAttr("shape", &shape));
+      shape_inference::ShapeHandle entry;
+      TF_RETURN_IF_ERROR(c->MakeShapeFromPartialTensorShape(shape, &entry));
+      c->set_output(0, entry);
+      return Status::OK();
+    });
 
 }  // namespace tensorflow
