+++ tensorflow_io/core/python/experimental/text_io_layer_ops.py
@@ -29,15 +29,16 @@ class TextIOLayer(tf.keras.layers.Layer):
   #=============================================================================
   def __init__(self, filename):
     """Obtain a text file IO layer to be used with tf.keras."""
-    self._resource = core_ops.io_layer_text_init(filename)
+    self._resource = core_ops.io_file_init(filename)
     super(TextIOLayer, self).__init__(trainable=False)
 
   def sync(self):
-    core_ops.io_layer_text_sync(self._resource)
+    core_ops.io_file_sync(self._resource)
 
   def call(self, inputs):
     content = tf.reshape(inputs, [tf.shape(inputs)[0], -1])
     if inputs.dtype != tf.string:
       content = tf.strings.as_string(content)
     content = tf.strings.reduce_join(content, axis=1, separator=',')
-    return core_ops.io_layer_text_call(inputs, content, self._resource)
+    content = content + tf.constant(["\n"])
+    return core_ops.io_file_call(content, False, self._resource)
